Use empty string to avoid loading GA script
In another episode of "this theme is not designed very well," apparently
leaving the `google_analytics_key` blank (or not listing it in
`_config.yml` at all) is not enough to disable it; presumably because it
parses to `None` instead of `""`. So we need to give the literal empty
string instead.

Note that to test this locally, you need to run as:

```bash
JEKYLL_ENV=production bundle exec jekyll serve
```

Without the `JEKYLL_ENV` argument set to `"production"`, you won't
generate the GA script loading in the version without this fix (so you
get a false positive that any attempt to fix this works).